api.js szétszedése, getLanguage áttétele apiAdminról apiMainre
